Created new file for helpers
Refactored and organized all helper functions into a new helper method.
