+++ api/server.js
@@ -9,6 +9,7 @@ const mysql = require("mysql2/promise"); //npm install mysql2
 const multer = require("multer"); //npm install multer
 const bcrypt = require("bcrypt"); //npm install bcrypt
 const cors = require("cors"); //npm install cors
+const helper = require("./helpers");  // adds helper methods from helpers.js
 
 const app = express();
 app.use(cors());
@@ -23,16 +24,6 @@ const CLIENT_ERROR_CODE_401 = 401;  // Unauthorized Access
 const SERVER_ERROR_CODE = 500;      // Server Error format: "An error has occured on the server!"
 const SERVER_ERROR_MESSAGE = "An error has occured on the server"
 const LOCAL_HOST = 8000;
-const DB_NAME = "huskyreads"; // Database name
-
-const db = mysql.createPool({
-	host: process.env.DB_URL || 'localhost',
-	port: process.env.DB_PORT || '3306',
-	user: process.env.DB_USERNAME || 'root',
-	password: process.env.DB_PASSWORD || 'root',
-	database: process.env.DB_NAME || DB_NAME,
-	multipleStatements: true
-});
 
 /* --------------------  ENDPOINTS  -------------------- */
 
@@ -59,7 +50,7 @@ app.post("/login", async (req, res) => {
 		if (!username || !password) {
 			res.status(CLIENT_ERROR_CODE_400).send("Missing username or password");
 		} else {
-			let returns = await getPassword(username);
+			let returns = await helper.getPassword(username);
 			if (returns.length >= 1 && password === returns[0]["password"]) {
 				res.status(SUCCESS_CODE).send("Login Successful");
 			} else {
@@ -82,11 +73,11 @@ app.post("/signup", async (req, res) => {
 		let password = req.body.password;
 		if (!username || !password) {
 			res.status(CLIENT_ERROR_CODE_400).send("Missing username or password");
-		} else if (await checkIfUsernameExists(username)) {
+		} else if (await helper.checkIfUsernameExists(username)) {
 			res.status(CLIENT_ERROR_CODE_400).send("Username already taken");
 		} else {
 			let info = [username, password];
-			await createUser(info);
+			await helper.createUser(info);
 			res.status(SUCCESS_CODE).send("Signup Successful");
 		}
 	} catch (err) {
@@ -105,14 +96,14 @@ app.post("/color_scheme", async (req, res) => {
 		let color_scheme = req.body.color_scheme;
 		if (!username || !color_scheme) {
 			res.status(CLIENT_ERROR_CODE_400).send("Missing username or color_scheme");
-		} else if (await !checkIfUsernameExists(usernamme)) {
+		} else if (await !helper.checkIfUsernameExists(usernamme)) {
 			res.status(CLIENT_ERROR_CODE_401).send("Invalid Username");
 		} else {
 			let info = [username, color_scheme];
 			if (!checkColor(color_scheme)) {
 				res.status(CLIENT_ERROR_CODE_400).send("Invalid Color Scheme");
 			} else {
-				await updateColorScheme(info);
+				await helper.updateColorScheme(info);
 				res.status(SUCCESS_CODE).send(SERVER_ERROR_MESSAGE);
 			}
 		}
@@ -133,7 +124,7 @@ app.get("/bookshelves/get/:username/:bookshelf", async function(req, res) {
 		let bookshelf = req.body.bookshelf;
 		if (!username) {
 			res.status(CLIENT_ERROR_CODE_400).send("Missing username paramter");
-		} else if (await !checkIfUsernameExists(username)) {
+		} else if (await !helper.checkIfUsernameExists(username)) {
 			res.status(CLIENT_ERROR_CODE_401).send("Invalid Username Parameter"); // Possibly change the error msg to something more clear?
 		} else {
 			if (!bookshelf) {
@@ -141,7 +132,7 @@ app.get("/bookshelves/get/:username/:bookshelf", async function(req, res) {
 			}
             // need to convert "All" into something meaningful here, or do in query
 			let info = [username, bookshelf];
-			let result = await getBook(info);   // getBook method needs working on
+			let result = await helper.getBook(info);   // getBook method needs working on
 
 			if (!result) {
 				res.status(CLIENT_ERROR_CODE_400).send("Invaild bookshelf name");
@@ -189,7 +180,7 @@ app.get("/books/search", async function(req, res) {
 		res.type("json");
 		let offset = req.query.offset ? parseInt(req.query.offset) : 0;
 		let resultLength = req.query.resultLength ? parseInt(req.query.resultLength) : 10;
-		let books = await getMatchingBooks(req.query);
+		let books = await helper.getMatchingBooks(req.query);
 		res.status(SUCCESS_CODE).json({
 			remainingBooksInSearch : books.slice(offset + resultLength).length,
 			books : books.slice(offset, offset + resultLength)
@@ -210,11 +201,11 @@ app.get("/books/detail/:isbn", async function(req, res) {
         if (!isbn) {
             res.status(CLIENT_ERROR_CODE_400).json({"error": "Missing ISBN Parameter"});
         } else {
-			let exists = await checkIfISBNExists(isbn);
+			let exists = await helper.checkIfISBNExists(isbn);
 			if (!exists) {
 				res.status(CLIENT_ERROR_CODE_400).json({"error": "Invalid ISBN"});
 			} else {
-				let results = await getBookDetails(isbn);
+				let results = await helper.getBookDetails(isbn);
 				res.json(results);
 			}
 		}
@@ -223,165 +214,6 @@ app.get("/books/detail/:isbn", async function(req, res) {
 		res.status(SERVER_ERROR_CODE).json({ err : SERVER_ERROR_MESSAGE });
 	}
 });
-
-/* -----------------  HELPER FUNCTIONS  ---------------- */
-
-// MISC FUNCTIONS
-/**
- * Creates new User based on info
- * @param {String[]} info
- */
- async function createUser(info) {
-	let query = "INSERT INTO User (username, password) VALUES (?, ?);";
-	await db.query(query, info);
-}
-
-/**
- * Updates the User's current Color Scheme.
- * @param {String[]} info
- */
-async function updateColorScheme(info) {
-	let query = "UPDATE User SET color_scheme = ? WHERE username = ?";
-	await db.query(query, info);
-}
-
-// CHECK FUNCTIONS
-/**
- * Checks if username exists
- * @param {String} username
- * @returns {boolean} true if username already exists
- */
-async function checkIfUsernameExists(username) {
-	let query = "SELECT * FROM User WHERE username = ?;";
-	let [rows] = await db.query(query, [username]);
-	return (rows.length >= 1);
-}
-
-/**
- * Checks if the given color scheme exists.
- * @param {String} color_scheme
- * @returns {boolean} true if the given color scheme exists
- */
-async function checkColor(color_scheme) {
-	let query = "SELECT * FROM User WHERE username = ? AND color_scheme = ?";
-	let [rows] = await db.query(query, info);
-	return (rows.length >= 1);
-}
-
-/**
- * @param {int} isbn Book ISBN number
- * @returns {boolean} True if a book exists with the given isbn, false otherwise
- * check if the book w/ the given isbn exists in our database
- */
- async function checkIfISBNExists(isbn) {
-    let query = "SELECT COUNT(*) AS count FROM Books WHERE Books.isbn = ?";
-    let [count] = await db.query(query, isbn);
-    return count[0].count > 0;
-}
-
-// GET FUNCTIONS
-/**
- * Gets password from username
- * @param {String} username
- * @returns info of username
- */
-async function getPassword(username) {
-	let query = "SELECT * FROM User WHERE username=?;";
-	let [rows] = await db.query(query, [username]);
-	return rows;
-}
-
-/**
- * Gets the book from the specified bookshelf
- * @param {String[]} info
- * @returns the Bookshelf information of the user.
- */
-async function getBook(info) {
-    // Need to fix this query
-    if (info[1] === "all") {
-        // do something meaningful with "all" parameter here
-    }
-    // need to update query
-	let query = "SELECT * FROM User WHERE username = ?;"; // Unsure how to address the other parts of the query.
-	let [rows] = await db.query(query, info);
-	return rows;
-}
-
-/**
- * Gets all of the books that match the given search query. Each book returned
- * will include its title, author(s) and isbn number.
- * @param {Object} info Search parameters provided by user
- * @returns {Object[]} List of books that satify the search query
- */
-async function getMatchingBooks(info) {
-	let title = info.title;
-	let author = info.author;
-	let genre = info.genre;
-
-	let params = [];
-	let query = `   DROP TEMPORARY TABLE IF EXISTS Results; CREATE TEMPORARY TABLE Results
-					SELECT Books.title AS title, Authors.name AS author_name, Books.ISBN AS isbn
-					FROM Books
-					INNER JOIN Book_Authors
-						ON Books.ISBN = Book_Authors.ISBN
-					INNER JOIN Authors
-						ON Book_Authors.id_author = Authors.id
-					INNER JOIN Book_Genre
-						ON Books.ISBN = Book_Genre.ISBN
-					INNER JOIN Genre
-						ON Book_Genre.id_genre = Genre.id
-					WHERE 1=1
-				`;
-	if (title) {
-		query += " AND Books.title = ?";
-		params.push(title);
-	}
-	if (genre) {
-		query += " AND Genre.name IN ?";
-		params.push([genre]);
-	}
-	query += `;
-		SELECT Results.title, GROUP_CONCAT(DISTINCT Results.author_name SEPARATOR ',') AS authors, Results.isbn
-		FROM Results
-		GROUP BY Results.title, Results.isbn
-	`;
-	if (author) {
-		query += "HAVING FIND_IN_SET(?, authors)";
-		params.push(author);
-	}
-	query += `;`;
-	return [await db.query(query, params)][0][0][2];
-}
-
-/**
- * @param {int} isbn Book ISBN number
- * @returns {JSON} json object containing the necessary book information (determined by API)
- */
- async function getBookDetails(isbn) {
-    let query = `
-    DROP TEMPORARY TABLE IF EXISTS Results;
-    CREATE TEMPORARY TABLE Results
-    SELECT Books.title AS title, Books.date_published AS date_published, Authors.name AS author_name, Genre.name AS genre_name
-    FROM Books
-    INNER JOIN Book_Authors
-        ON Books.ISBN = Book_Authors.ISBN
-    INNER JOIN Authors
-        ON Book_Authors.id_author = Authors.id
-    INNER JOIN Book_Genre
-        ON Books.ISBN = Book_Genre.ISBN
-    INNER JOIN Genre
-        ON Book_Genre.id_genre = Genre.id
-    WHERE Books.ISBN = ?
-    ;
-    SELECT Results.title, Results.date_published, GROUP_CONCAT(DISTINCT Results.author_name SEPARATOR ',') AS authors, GROUP_CONCAT(DISTINCT Results.genre_name SEPARATOR ',') AS genres
-    FROM Results
-    GROUP BY Results.title, Results.date_published
-    ;
-    `
-    let [results] = await db.query(query, isbn);
-    return results[2];
-}
-
 
 /* ------------------  LOGGING MODULE  ----------------- */
 /**
